bug-hunt #115: memory manager - fix -Wconversion sign-conversion warnings
Fix 16 -Wconversion (sign-conversion) warnings across 7 memory manager files:

slab.c (6):
- Cast int total_in_cache/free_in_cache to uint64_t for slab_stats
- Cast int j to uint64_t in pmm_free_frame loop
- Cast int cache->num to size_t in kmalloc sizeof multiply
- Cast int i to size_t in pointer arithmetic with size_t

zswap.c (3):
- Cast int e->comp_len to uint64_t for zswap_pool_used subtraction
- Cast int len to uint64_t in writeback path

zsmalloc.c (2):
- Cast int size field to size_t in zsmalloc_alloc
- Cast int division result to uintptr_t in zsmalloc_free

ksm.c (1):
- Cast int scanned to uint64_t for ksm_total_scanned

cma.c (2):
- Cast int bitmap_size to size_t for pmm_alloc_frames and memset

hugepage_migration.c (4):
- Cast int nr_pages to uint64_t for isolated page counters

zbud.c (1):
- Cast int chunk_idx to uint64_t in handle construction

+++ src/memory/cma.c
@@ -26,10 +26,10 @@ int cma_create_area(uint64_t base_pfn, uint64_t count, const char *name) {
 
     /* Allocate bitmap: one bit per page */
     area->bitmap_size = (int)((count + 7) / 8);
-    area->bitmap = (uint64_t *)pmm_alloc_frames((area->bitmap_size + PAGE_SIZE - 1) / PAGE_SIZE);
+    area->bitmap = (uint64_t *)pmm_alloc_frames((size_t)(area->bitmap_size + PAGE_SIZE - 1) / PAGE_SIZE);
     if (!area->bitmap) return -ENOMEM;
 
-    memset(area->bitmap, 0, area->bitmap_size);
+    memset(area->bitmap, 0, (size_t)area->bitmap_size);
     strncpy(area->name, name, sizeof(area->name) - 1);
     area->name[sizeof(area->name) - 1] = '\0';
     area->initialized = 1;
